style: Add missing semi colons in JS files and add some ESLint rules to pass undef and unused vars
JS code in this projects needs some refactoring. It seems like a good way to extract code to modules. So we mark this task as a TODO, but now to work with project I make some style fixes and add rules for ESLint to pass undef objects and unused-vars as a temporary decision.